Forward port from BRANCH
svn path=/trunk/kdenetwork/kopete/; revision=307594

+++ protocols/yahoo/libyahoo2/sha.c
@@ -50,7 +50,7 @@ void shaUpdate(SHA_CTX *ctx, unsigned char *dataIn, int len) {
    */
   for (i = 0; i < len; i++) {
     ctx->W[ctx->lenW / 4] <<= 8;
-    ctx->W[ctx->lenW / 4] |= (unsigned long)dataIn[i];
+    ctx->W[ctx->lenW / 4] |= (uint32)dataIn[i];
     if ((++ctx->lenW) % 64 == 0) {
       shaHashBlock(ctx);
       ctx->lenW = 0;
@@ -109,7 +109,7 @@ void shaBlock(unsigned char *dataIn, int len, unsigned char hashout[20]) {
 
 static void shaHashBlock(SHA_CTX *ctx) {
   int t;
-  unsigned long A,B,C,D,E,TEMP;
+  uint32 A,B,C,D,E,TEMP;
 
   for (t = 16; t <= 79; t++)
     ctx->W[t] =
